+++ src/util/zs_strutils.h
@@ -88,7 +88,7 @@ namespace zetscript{
 		bool contains(const std::string & input, const std::string & str_containts,StringComparer sc=StringComparer::Exactly);
 		bool contains(const std::vector<std::string> & input, const std::string & str_containts,StringComparer sc=StringComparer::Exactly);
 
-		int index_of(std::string& text, std::string& pattern);
+		size_t index_of(std::string& text, std::string& pattern);
 
 		std::string unescape(const std::string & s);
 	}
